Render the documentation as a website with MkDocs
MkDocs is an open source Python library for generating documentation
from Markdown files. Since you have some really nice documentation I
thought this might be a valuable addition - particularly if you want to
host the documentation somewhere or put it on ReadTheDocs (they support
MkDocs).

The only requirement MkDocs has is that the home page is named index.md
which is why I have named that. Otherwise it works great.

Note, I have not added MkDocs to the requirements, but you will
obviously need to install it to try it out. so, `pip install mkdocs` and
then `mkdocs serve` and go to localhost:8000 in your browser. To build
the documentation to be used somewhere else use the command `mkdocs
build` which will output the documentation to a `site` directory